Added My Reviews Page & Worked on favourite functionality
Added a page that displays the users reviews, allowing them to edit or delete said reviews. I also started working on adding a favourite button to each of the restaurants, working on adding a double click or a single click for adding and removing from the list. Fixed a few buys and tidied up some other code. Started working later today due to a migraine but still a positive day, got a few end points sorted :)