Updated Extension Tasks
Implmented new page dedicated to extension task on distance of locations - you can now find your nearest cafe on the click of a button. Added a new tab option with this page, links to feed page if you want to find out more about the location that is closest. Also attempted to fix a new bug with no luck, pages flickering on load. Will try to fix this before the deadline.