Corrected Styleguide
Installed Eslint and applied the airbnb style guide to my code, spent the day going through (most) of my pages in order and cleaned the code, to the best of my understanding. Will do the remaining 4 pages tomorrow. Also used this as an opertunity to fix outstanding bugs, and cleaned up all api requests, to make the feedback from the server as clear as possible to both the user and any future programmers reading code